Protected item routes

+++ Backend/Controllers/itemController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Item = require('../models/ItemModel')
+const User = require('../Models/UserModel')
 
 // @desc    Get Items
 // @route   GET /api/items
 // @access  Private
 const getItems = asyncHandler(async (req, res) => {
-    const items = await Item.find()
+    const items = await Item.find({ user: req.user.id })
 
     res.status(200).json(items)
 })
@@ -21,7 +22,8 @@ const setItems = asyncHandler(async (req, res) => {
     }
 
     const item = await Item.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(item)
@@ -38,6 +40,18 @@ const updateItems = asyncHandler(async (req, res) => {
         throw new Error('Item not found!')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (item.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedItem = await Item.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json(updatedItem)
@@ -52,6 +66,18 @@ const deleteItems = asyncHandler(async (req, res) => {
     if (!item) {
         res.status(400)
         throw new Error('Item not found!')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (item.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await item.deleteOne()
